Move code out of `__init__` file
For stylistic reasons I think it's nicer not to have substantive code in
`__init__` files and to have them for the most part just re-import names
from elsewhere.
